Fix token validation

+++ app/Http/Controllers/TeamController.php
@@ -16,7 +16,8 @@ class TeamController extends Controller
      *
      * @return void
      */
-    public static $owner_role_id,$member_role_id;
+    public static $owner_role_id, $member_role_id;
+
     public function __construct()
     {
         //
@@ -48,7 +49,6 @@ class TeamController extends Controller
         $user = User::where('token', $token)->first();
         if ($token && $user) {
             $user_id = $user->id;
-
             $data = [
                 'title' => $request['title'],
             ];
@@ -61,11 +61,12 @@ class TeamController extends Controller
                 return response()->json($validator->messages(), 200);
             }
             $team = Team::create($data);
-            self::$owner_role_id = Role::where('status','Owner')->first()->id;
-            $owner_role_id =  self::$owner_role_id;
+            self::$owner_role_id = Role::where('status', 'Owner')->first()->id;
+            $owner_role_id = self::$owner_role_id;
+
             UserRoleTeam::create([
                 'user_id' => $user_id,
-                'tam_id' => $team->id,
+                'team_id' => $team->id,
                 'role_id' => $owner_role_id,
             ]);
 
@@ -81,7 +82,9 @@ class TeamController extends Controller
     {
         $token = $request->headers->all()['token'][0];
         $user = User::where('token', $token)->first();
-        if($token && $user){
+        $team_owner = UserRoleTeam::where([['user_id', $user->id], ['team_id', $id]])->first();
+        if ($token && $user) {
+
             $data = [
                 'title' => $request['title'],
             ];
@@ -92,80 +95,103 @@ class TeamController extends Controller
             if ($validator->fails()) {
                 return response()->json($validator->messages(), 200);
             } else {
-                Team::where('id', $id)->update($data);
-                return response()->json(['status' => 'success', 'message' => 'Team updated'], 200);
+                if ($team_owner) {
+                    Team::where('id', $id)->update($data);
+                    return response()->json(['status' => 'success', 'message' => 'Team updated'], 200);
+                } else {
+                    return response()->json(['status' => 'failed', 'message' => 'You dont have an access to update the team'], 200);
+                }
+
             }
-        }
-        else{
-            return response()->json(['status' => 'failed', 'message' =>'Invalid token or Invalid user']);
+        } else {
+            return response()->json(['status' => 'failed', 'message' => 'Invalid token or Invalid user']);
         }
 
     }
 
 
-    public function delete(Request $request,$id)
+    public function delete(Request $request, $id)
     {
         $token = $request->headers->all()['token'][0];
         $user = User::where('token', $token)->first();
-        if($token && $user){
+        $team_owner = UserRoleTeam::where([['user_id', $user->id], ['team_id', $id]])->first();
+        if ($token && $user) {
+
             $team = Team::where('id', $id)->first();
             if (!empty($team)) {
-                $team->delete();
-                return response()->json(['status' => 'success', 'message' => 'Team successfully deleted'], 200);
+                if ($team_owner) {
+                    $team->delete();
+                    return response()->json(['status' => 'success', 'message' => 'Team successfully deleted'], 200);
+                } else {
+                    return response()->json(['status' => 'success', 'message' => 'You dont have an access to delete the team'], 200);
+                }
             } else {
                 return response()->json(['status' => 'failed', 'message' => 'Team not deleted'], 200);
             }
-        }
-        else{
-            return response()->json(['status' => 'failed', 'message' =>'Invalid token or Invalid user']);
+        } else {
+            return response()->json(['status' => 'failed', 'message' => 'Invalid token or Invalid user']);
         }
 
     }
 
     public function addTeamMember(Request $request, $member_id, $team_id)
     {
-        self::$member_role_id = Role::where('status','Member')->first()->id;
+        self::$member_role_id = Role::where('status', 'Member')->first()->id;
         $member_role_id = self::$member_role_id;
         $token = $request->headers->all()['token'][0];
         $user = User::where('token', $token)->first();
+        $team_owner = UserRoleTeam::where([['user_id', $user->id], ['team_id', $team_id]])->first();
         if ($token && $user) {
-            $team_member_info = UserRoleTeam::where([['user_id', $member_id], ['team_id', $team_id]])->first();
-            if (!empty($team_member_info)) {
-                UserRoleTeam::where('user_id', $member_id)->update(['role_id' => $member_role_id]);
-                return response()->json(['status' => 'success', 'message' => 'Team member  updated'], 200);
-            } else {
-                UserRoleTeam::create([
-                    'user_id' => $member_id,
-                    'team_id' => $team_id,
-                    'role_id' => $member_role_id,
-                ]);
+            if ($team_owner) {
+                $team_member_info = UserRoleTeam::where([['user_id', $member_id], ['team_id', $team_id]])->first();
+                if (!empty($team_member_info)) {
+                    UserRoleTeam::where('user_id', $member_id)->update(['role_id' => $member_role_id]);
+                    return response()->json(['status' => 'success', 'message' => 'Team member  updated'], 200);
+                } else {
+                    UserRoleTeam::create([
+                        'user_id' => $member_id,
+                        'team_id' => $team_id,
+                        'role_id' => $member_role_id,
+                    ]);
 
+                }
+                return response()->json(['status' => 'success', 'message' => 'Team member added'], 200);
+            } else {
+                return response()->json(['status' => 'success', 'message' => 'You dont have an access to delete the team'], 200);
             }
-            return response()->json(['status' => 'success', 'message' => 'Team member added'], 200);
         } else {
             return response()->json(['status' => 'failed', 'message' => 'Invalid token or Invalid user'], 200);
         }
+
     }
+
 
     public function addTeamOwner(Request $request, $owner_id, $team_id)
     {
         $owner_role_id = self::$owner_role_id;
         $token = $request->headers->all()['token'][0];
         $user = User::where('token', $token)->first();
+        $team_owner = UserRoleTeam::where([['user_id', $user->id], ['team_id', $team_id]])->first();
         if ($token && $user) {
-            $team_owner_info = UserRoleTeam::where([['user_id', $owner_id], ['team_id', $team_id]])->first();
-            if (!empty($team_owner_info)) {
-                UserRoleTeam::where('user_id', $owner_id)->update(['role_id' => $owner_role_id]);
-                return response()->json(['status' => 'success', 'message' => 'Team owner updated'], 200);
-            } else {
-                UserRoleTeam::create([
-                    'user_id' => $owner_id,
-                    'team_id' => $team_id,
-                    'role_id' => $owner_role_id,
-                ]);
+            if ($team_owner) {
+                $team_owner_info = UserRoleTeam::where([['user_id', $owner_id], ['team_id', $team_id]])->first();
+                if (!empty($team_owner_info)) {
+                    UserRoleTeam::where('user_id', $owner_id)->update(['role_id' => $owner_role_id]);
+                    return response()->json(['status' => 'success', 'message' => 'Team owner updated'], 200);
+                } else {
+                    UserRoleTeam::create([
+                        'user_id' => $owner_id,
+                        'team_id' => $team_id,
+                        'role_id' => $owner_role_id,
+                    ]);
 
-                return response()->json(['status' => 'success', 'message' => 'Team owner added'], 200);
+                    return response()->json(['status' => 'success', 'message' => 'Team owner added'], 200);
+                }
+
+            } else {
+                return response()->json(['status' => 'success', 'message' => 'You dont have an access to delete the team'], 200);
             }
+
         } else {
             return response()->json(['status' => 'failed', 'message' => 'Invalid token or Invalid user'], 200);
         }
@@ -175,21 +201,27 @@ class TeamController extends Controller
     public function deleteTeamMember(Request $request, $team_id, $member_id)
     {
         $token = $request->headers->all()['token'][0];
-        $user = User::where('token', $token)->get();
+        $user = User::where('token', $token)->first();
+        $team_owner = UserRoleTeam::where([['user_id', $user->id], ['team_id', $team_id]])->first();
         if ($token && $user) {
-            if ($team_id) {
-                if ($member_id) {
-                    $member = UserRoleTeam::where('user_id', $member_id)->first();
-                    if (!empty($member)) {
-                        UserRoleTeam::where('user_id', $member_id)->delete();
-                        return response()->json(['status' => 'success', 'message' => 'Member deleted'], 200);
-                    } else {
-                        return response()->json(['status' => 'failed', 'message' => 'Member not found'], 200);
+            if ($team_owner) {
+                if ($team_id) {
+                    if ($member_id) {
+                        $member = UserRoleTeam::where('user_id', $member_id)->first();
+                        if (!empty($member)) {
+                            UserRoleTeam::where('user_id', $member_id)->delete();
+                            return response()->json(['status' => 'success', 'message' => 'Member deleted'], 200);
+                        } else {
+                            return response()->json(['status' => 'failed', 'message' => 'Member not found'], 200);
+                        }
                     }
+                } else {
+                    return response()->json(['status' => 'failed', 'message' => 'Team not found'], 200);
                 }
             } else {
-                return response()->json(['status' => 'failed', 'message' => 'Team not found'], 200);
+                return response()->json(['status' => 'success', 'message' => 'You dont have an access to delete the team'], 200);
             }
+
         } else {
             return response()->json(['status' => 'failed', 'message' => 'Invalid token or Invalid user'], 200);
         }
@@ -201,19 +233,27 @@ class TeamController extends Controller
     {
         $token = $request->headers->all()['token'][0];
         $user = User::where('token', $token)->first();
+        $team_owner = UserRoleTeam::where([['user_id', $user->id], ['team_id', $team_id]])->first();
         if ($token && $user) {
-            if ($team_id) {
-                if ($owner_id) {
-                    $owner = UserRoleTeam::where('user_id', $owner_id)->first();
-                    if (!empty($owner)) {
-                        UserRoleTeam::where('user_id', $owner_id)->delete();
-                        return response()->json(['status' => 'success', 'message' => 'Owner deleted'], 200);
+            if ($team_owner) {
+
+                if ($team_id) {
+                    if ($owner_id) {
+                        $owner = UserRoleTeam::where('user_id', $owner_id)->first();
+                        if (!empty($owner)) {
+                            UserRoleTeam::where('user_id', $owner_id)->delete();
+                            return response()->json(['status' => 'success', 'message' => 'Owner deleted'], 200);
+                        } else {
+                            return response()->json(['status' => 'failed', 'message' => 'Owner not found'], 200);
+                        }
                     } else {
-                        return response()->json(['status' => 'failed', 'message' => 'Owner not found'], 200);
+                        return response()->json(['status' => 'failed', 'message' => 'Team not found'], 200);
                     }
-                } else {
-                    return response()->json(['status' => 'failed', 'message' => 'Team not found'], 200);
                 }
+
+            } else {
+                return response()->json(['status' => 'success', 'message' => 'You dont have an access to delete the team'], 200);
+
             }
         } else {
             return response()->json(['status' => 'failed', 'message' => 'Invalid token or Invalid user']);
